users list dialog update

+++ app/src/main/java/com/example/ecoapp/presentation/fragments/UserListDialogFragment.java
@@ -1,17 +1,19 @@
 package com.example.ecoapp.presentation.fragments;
 
-import android.app.AlertDialog;
-import android.app.Dialog;
 import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.databinding.DataBindingUtil;
 import androidx.fragment.app.DialogFragment;
 import androidx.lifecycle.ViewModelProvider;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.ecoapp.R;
 import com.example.ecoapp.data.models.User;
+import com.example.ecoapp.databinding.UsersListDialogBinding;
 import com.example.ecoapp.domain.helpers.StorageHandler;
 import com.example.ecoapp.presentation.adapters.UserScoresAdapter;
 import com.example.ecoapp.presentation.viewmodels.EventViewModel;
@@ -20,26 +22,30 @@ import com.example.ecoapp.presentation.viewmodels.ProfileViewModel;
 import java.util.ArrayList;
 
 public class UserListDialogFragment extends DialogFragment {
+    private UsersListDialogBinding binding;
     public UserScoresAdapter adapter;
     public EventViewModel viewModel;
     public ProfileViewModel profileViewModel;
     public ArrayList<User> users;
     public String eventID;
     public int theme;
-    public RecyclerView recyclerView;
 
-    @NonNull
+    @Nullable
     @Override
-    public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        binding = DataBindingUtil.inflate(inflater, R.layout.users_list_dialog, container, false);
+        theme = new StorageHandler(requireContext()).getTheme();
+
+        binding.setThemeInfo(theme);
+        if (getDialog() != null) getDialog().setTitle("Cписок пользователей");
+
         viewModel = new ViewModelProvider(this).get(EventViewModel.class);
         profileViewModel = new ViewModelProvider(this).get(ProfileViewModel.class);
         Bundle args = getArguments();
 
-        theme = new StorageHandler(requireContext()).getTheme();
 
-        recyclerView = new RecyclerView(requireContext());
-        recyclerView.setHasFixedSize(true);
-        recyclerView.setLayoutManager(new LinearLayoutManager(requireContext(), LinearLayoutManager.VERTICAL, false));
+        binding.usersListRecycler.setHasFixedSize(true);
+//        binding.usersListRecycler.setLayoutManager(new LinearLayoutManager(requireContext(), LinearLayoutManager.VERTICAL, false));
 
         if (args != null) {
             eventID = args.getString("eventID");
@@ -48,25 +54,37 @@ public class UserListDialogFragment extends DialogFragment {
                     this.users = users;
 
                     adapter = new UserScoresAdapter(users, theme);
-                    recyclerView.setAdapter(adapter);
+                    binding.usersListRecycler.setAdapter(adapter);
 
-                    adapter.setOnItemClickListener((position) -> {
-                        profileViewModel.updateUserScores(users.get(position).getId(), eventID).observe(requireActivity(), statusCode -> {
-                            if (statusCode >= 200 && statusCode < 400) {
-                                adapter.deleteItem(users.get(position).getId());
-                            }
-                        });
-                    });
+//                    adapter.setOnItemClickListener((position) -> {
+//                        profileViewModel.updateUserScores(users.get(position).getId(), eventID).observe(requireActivity(), statusCode -> {
+//                            if (statusCode >= 200 && statusCode < 400) {
+//                                adapter.deleteItem(users.get(position).getId());
+//                            }
+//                        });
+//                    });
                 }
             });
         }
 
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        builder.setTitle("Список пользователей").setView(recyclerView);
+        return binding.getRoot();
+    }
 
-        AlertDialog dialog = builder.create();
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
 
-        return dialog;
+        viewModel.getIsContext().observe(getViewLifecycleOwner(), isCtx -> {
+            if (isCtx) {
+                adapter.setOnItemClickListener((position) -> {
+                    profileViewModel.updateUserScores(users.get(position).getId(), eventID).observe(requireActivity(), statusCode -> {
+                        if (statusCode >= 200 && statusCode < 400) {
+                            adapter.deleteItem(users.get(position).getId());
+                        }
+                    });
+                });
+            }
+        });
     }
 
     @Override
